Update nginx to build with conf file
This updates the nginx repository to include a Dockerfile with
instructions to build an nginx image with our nginx.conf file
copied over to the image rather than using it from the host. This
is because we are no longer pulling source code down to the
webserver. The deploy process only pulls images from the Docker
registry.