Run nginx as non-root and logs
This updates the nginx Dockerfile to run the image as a non-root
user for security reasons. Since we are running as non-root, we
cannot bind to port 80 in the container, so we'll need to use 8080
instead. That's fine since the docker-compose.yml configuration
will be updated to map 8080 from the container to 80 on the
webserver.

We also update the log files to prefix them with nginx-... since
they will be writing to the same webserver directory as the other
services.

The pid location also has to be updated to get stored in a
directory that the non-root user has write access to. /tmp is
a perfect location for this.

And finally, we update the nginx image from 1.27.2 to 1.27.3.