Update dockerfile to override the error file
nginx will always create an error_log at /var/log/nginx/error.log.
This isn't good for us because the /var/log/nginx directory is
mounted to the deploy servers ./log directory. In this log
directory are all service's logs, so we prefix all the logs to be
able to distingush between the different service's logs. When nginx
creates error.log, it's missing the prefix. Even adding a
configuration to nginx.conf for error_log isn't respected since
this is apparently compiled into the binary. The only way to
override this log file is the -e flag.