Lock down nginx conf
This configures less generous timeouts for nginx clients and proxy
requests to our services. We also add request rate limits and
concurrent limits.

We also change the URL domains to our registered URLs.

We still don't have SSL configured because we are hosting this
application internally on our local LAN. The tincan subdomains are
configured to point to internal private RFC1918 IP addresses. Until
we obtain a public server, this will be the only way to access the
website.